requirements.txt
python-dotenv
python-telegram-bot
openai>=1.0.0
discord